change dbdir_tmp, improve if statements, remove comments

+++ root/scripts/init.sh
@@ -1,10 +1,11 @@
 #!/bin/bash
 
-set -ex
+set -e
 
 dbdir="/config/Plex Media Server/Plug-in Support/Databases"
-dbdir_tmp="/tmp/plex/databases"
+dbdir_tmp="/plex-db"
 dbdir_backup="/config/backup-databases"
+dbdir_parent="/config/Plex Media Server/Plug-in Support"
 
 mkdir -p "$dbdir_backup"
 mkdir -p "cachedir_tmp"
@@ -17,14 +18,17 @@ if [ ! -d "$dbdir_backup" ]; then
 fi
 
 # database should only be copied to tmpfs if it is empty
-if [[ ( ! -d "$dbdir_tmp" ) || ( ! "$(ls -A $dbdir_tmp)" ) ]]; then
+if [[ ( ! -d "$dbdir_tmp" ) || ( -d "$dbdir_tmp" && ! "$(ls -A $dbdir_tmp)" ) ]]; then
   echo "Copying database to tmpfs..."
   rsync -acz --delete "$dbdir_backup/" "$dbdir_tmp"
   echo "Database copied to tmpfs"
 fi
 
-if [[ ! -L "$dbdir" ]]; then
-  rm -r "$dbdir"
+if [[ ! "$(find "$dbdir_parent" -lname $dbdir_tmp)" ]]; then
+  if [[ -f "$dbdir" ]]; then
+    rm -r "$dbdir"
+  fi
+
   ln -s "$dbdir_tmp" "$dbdir"
 fi
 
@@ -40,7 +44,7 @@ function backup() {
 trap 'backup' SIGINT SIGTERM
 
 # set cronjob, this can probably be done in dockerfile
-crontab /scripts/cron-backup
+#crontab /scripts/cron-backup
 
 # start command
 runuser -u nobody -- /home/nobody/start.sh &
